third_party/WebKit/LayoutTests/http/tests/serviceworker/chromium.fetch-event-request-body.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- This is a temporary test and should be deleted once the WPT
  test fetch-event.https.html is passing in Chromium. It's only needed to
  allow test coverage while that test is failing. -->
<title>FetchEvent#request.body</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
// This tests that FetchEvent#request.body is set. The service
// worker reads the body and then postMessage()'s back to the
// main page. It's using postMessage since there is some other
// bug with S13nSW that prevents using respondWith() with POST.
promise_test(t => {
   const kScope = 'resources/new';
   const kScript = 'resources/fetch-event-request-body-worker.js';
   let sw;

   return service_worker_unregister_and_register(t, kScript, kScope)
     .then(reg => {
         add_completion_callback(() => {
             reg.unregister();
           });
         sw = reg.installing;
         return wait_for_state(t, sw, 'activated');
       })
     .then(() => {
         return with_iframe(kScope);
       })
     .then(frame => {
         add_completion_callback(() => {
             frame.remove();
           });

         sw.postMessage('ping');
         frame.contentWindow.fetch('blank.html?fetch-with-body', {
             method: 'POST',
             body: 'i am the request body'
           });
         return new Promise(resolve => {
             navigator.serviceWorker.addEventListener('message', e => {
                 resolve(e.data);
               });
           });
       })
     .then(data => {
         assert_equals(data, 'i am the request body');
       });
  }, 'FetchEvent#request.body is set');
</script>
